requirements.txt
tensorflow-cpu==2.5.0
Keras==2.4.3
numpy==1.19.5
streamlit==0.82.0
Pillow==8.2.0
